requirements.txt
python-telegram-bot==21.6
openai==1.52.0
python-docx==1.1.2
python-dotenv==1.0.1
httpx==0.27.2
